refactor: move networkHrp store to profile packages
