+++ packages/desktop/components/popups/MintNftFormPopup.svelte
@@ -5,7 +5,7 @@
     import { BaseError } from '@core/error/classes'
     import { handleError } from '@core/error/handlers/handleError'
     import { localize } from '@core/i18n'
-    import { networkHrp } from '@core/network/stores'
+    import { networkHrp } from '@core/profile/stores'
     import { MimeType } from '@core/nfts/types'
     import { isValidUri } from '@core/utils/validation'
     import { validateBech32Address } from '@core/utils/crypto'
