+++ packages/shared/components/inputs/AliasInput.svelte
@@ -2,7 +2,7 @@
     import { Modal, SelectorInput, IOption } from 'shared/components'
     import { selectedAccount } from '@core/account/stores'
     import { localize } from '@core/i18n'
-    import { networkHrp } from '@core/network/stores'
+    import { networkHrp } from '@core/profile/stores'
     import { validateBech32Address } from '@core/utils/crypto'
     import { ADDRESS_TYPE_ALIAS } from '@core/wallet/constants'
     import { convertHexAddressToBech32 } from '@core/wallet/utils'
