+++ packages/shared/components/inputs/RecipientInput.svelte
@@ -3,8 +3,7 @@
     import { selectedAccountIndex } from '@core/account/stores'
     import { getAccountColorById } from '@core/account/utils'
     import { localize } from '@core/i18n'
-    import { networkHrp } from '@core/network/stores'
-    import { visibleActiveAccounts } from '@core/profile/stores'
+    import { visibleActiveAccounts, networkHrp } from '@core/profile/stores'
     import { validateBech32Address, validateEthereumAddress } from '@core/utils/crypto'
     import { Subject } from '@core/wallet/types'
     import { getSubjectFromAddress } from '@core/wallet/utils'
