+++ packages/shared/lib/core/network/stores/index.ts
@@ -1,3 +1,2 @@
-export * from './network-hrp.store'
 export * from './network-status.store'
 export * from './node-info.store'
