+++ packages/shared/lib/core/profile/stores/index.ts
@@ -6,6 +6,7 @@ export * from './is-software-profile.store'
 export * from './is-destroying-manager.store'
 export * from './is-active-ledger-profile.store'
 export * from './login-progress.store'
+export * from './network-hrp.store'
 export * from './profiles.store'
 export * from './timeSinceStrongholdUnlocked.store'
 export * from './timeStrongholdLastUnlocked.store'
